CU-8692xcr3r: Add some abstraction for testing (regarding spacy model)
